added feature of autopplay, complete lectures, complete sections
added code to track now playing which now triggers the current video. the now playing change triggers another function which checks if the sections is expanded or not which is handly on autoplay and after the lectures are loaded on dom brings the now playing lecture to scrollintoview. added support to toogle the lecture completed and keep a record of completed lectures, if all the lectures are checked completed on the section the section is automatically gets checked completed. added event to track onvideoend which  marks the current lecturte completed and checks if autoplay is on plays next video. added code to create a dictionary which stores lectureid with url, next and previous lecture id. and also added code to disable previous and next button on first and last lecture of course